+++ AdminProject/app/src/main/java/org/admin/repository/TextReportDao.java
@@ -11,10 +11,10 @@ public interface TextReportDao {
     public List<Report> findAllBy(String type);
 
     public Report findBy(@Param("type") String type,
-                         @Param("no") int no,
-                         @Param("memberNo") int memberNo);
+                         @Param("no") long no,
+                         @Param("memberNo") long memberNo);
 
-    public int updateState(@Param("reportNo") int reportNo);
+    public int updateState(@Param("reportNo") long reportNo);
 
     public int updateBoardState(@Param("report") Report report, @Param("boardState") String boardState);
 }
