+++ AdminProject/app/src/main/java/org/admin/service/TextReportService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface TextReportService {
     public List<Report> getAllBy(String type);
 
-    public Report getBy(String type, int targetNo, int memberNo);
+    public Report getBy(String type, long targetNo, long memberNo);
 
-    public int updateState(int reportNo);
+    public int updateState(long reportNo);
 
     public int updateBoardState(Report report, String boardState);
 }
